fix(workspaces): transactions in events
