+++ packages/server/modules/workspaces/index.ts
@@ -6,30 +6,8 @@ import { Optional, SpeckleModule } from '@/modules/shared/helpers/typeHelper'
 import { workspaceRoles } from '@/modules/workspaces/roles'
 import { workspaceScopes } from '@/modules/workspaces/scopes'
 import { registerOrUpdateRole } from '@/modules/shared/repositories/roles'
-import {
-  initializeEventListenersFactory,
-  onInviteFinalizedFactory,
-  onProjectCreatedFactory,
-  onWorkspaceRoleDeletedFactory,
-  onWorkspaceRoleUpdatedFactory
-} from '@/modules/workspaces/events/eventListener'
-import {
-  getWorkspaceRolesFactory,
-  getWorkspaceWithDomainsFactory,
-  upsertWorkspaceRoleFactory
-} from '@/modules/workspaces/repositories/workspaces'
-import {
-  deleteProjectRoleFactory,
-  getStream,
-  upsertProjectRoleFactory
-} from '@/modules/core/repositories/streams'
-import { updateWorkspaceRoleFactory } from '@/modules/workspaces/services/management'
-import { getEventBus } from '@/modules/shared/services/eventBus'
-import { getStreams } from '@/modules/core/services/streams'
-import { findVerifiedEmailsByUserIdFactory } from '@/modules/core/repositories/userEmails'
+import { initializeEventListenersFactory } from '@/modules/workspaces/events/eventListener'
 import { validateModuleLicense } from '@/modules/gatekeeper/services/validateLicense'
-import { queryAllWorkspaceProjectsFactory } from '@/modules/workspaces/services/projects'
-import { mapWorkspaceRoleToInitialProjectRole } from '@/modules/workspaces/domain/logic'
 
 const { FF_WORKSPACES_MODULE_ENABLED } = getFeatureFlags()
 
@@ -59,34 +37,7 @@ const workspacesModule: SpeckleModule = {
     moduleLogger.info('⚒️  Init workspaces module')
 
     if (isInitial) {
-      quitListeners = initializeEventListenersFactory({
-        onProjectCreated: onProjectCreatedFactory({
-          getDefaultWorkspaceProjectRoleMapping: mapWorkspaceRoleToInitialProjectRole,
-          upsertProjectRole: upsertProjectRoleFactory({ db }),
-          getWorkspaceRoles: getWorkspaceRolesFactory({ db })
-        }),
-        onWorkspaceRoleDeleted: onWorkspaceRoleDeletedFactory({
-          queryAllWorkspaceProjects: queryAllWorkspaceProjectsFactory({ getStreams }),
-          deleteProjectRole: deleteProjectRoleFactory({ db })
-        }),
-        onWorkspaceRoleUpdated: onWorkspaceRoleUpdatedFactory({
-          getDefaultWorkspaceProjectRoleMapping: mapWorkspaceRoleToInitialProjectRole,
-          queryAllWorkspaceProjects: queryAllWorkspaceProjectsFactory({ getStreams }),
-          deleteProjectRole: deleteProjectRoleFactory({ db }),
-          upsertProjectRole: upsertProjectRoleFactory({ db })
-        }),
-        onInviteFinalized: onInviteFinalizedFactory({
-          getStream,
-          logger: moduleLogger,
-          updateWorkspaceRole: updateWorkspaceRoleFactory({
-            getWorkspaceWithDomains: getWorkspaceWithDomainsFactory({ db }),
-            findVerifiedEmailsByUserId: findVerifiedEmailsByUserIdFactory({ db }),
-            getWorkspaceRoles: getWorkspaceRolesFactory({ db }),
-            upsertWorkspaceRole: upsertWorkspaceRoleFactory({ db }),
-            emitWorkspaceEvent: (...args) => getEventBus().emit(...args)
-          })
-        })
-      })()
+      quitListeners = initializeEventListenersFactory({ db })()
     }
     await Promise.all([initScopes(), initRoles()])
   },
